% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha transparency to hex colors}
\usage{
add_alpha(hex, alpha = 1)
}
\arguments{
\item{hex}{Character vector of hex colors}

\item{alpha}{Numeric value between 0 (transparent) and 1 (opaque)}
}
\value{
Character vector of hex colors with alpha channel
}
\description{
Add alpha transparency to hex colors
}
\examples{
add_alpha("#FF0000", 0.5)  # Semi-transparent red
add_alpha(c("#FF0000", "#00FF00"), 0.7)

}
