% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example_microbiome}
\alias{example_microbiome}
\title{Example microbiome dataset}
\format{
A data frame with columns:
\describe{
\item{SampleID}{Sample identifier}
\item{OTU}{ASV/OTU identifier}
\item{Counts}{Raw read counts}
\item{Depth}{Total sequencing depth per sample}
\item{RA}{Relative abundance (Counts / Depth)}
\item{Habitat}{Broad habitat category (Terrestrial, Oceanic, Freshwater, Brackish)}
\item{Description}{Sample description)}
\item{SampleType}{Original sample type label}
\item{Kingdom}{Taxonomic kingdom}
\item{Phylum}{Taxonomic phylum}
\item{Class}{Taxonomic class}
\item{Order}{Taxonomic order}
\item{Family}{Taxonomic family}
\item{Genus}{Taxonomic genus}
}

@source Caporaso, J.G., et al. (2011). Global patterns of 16S rRNA
diversity at a depth of millions of sequences per sample.
\emph{PNAS}, 108, 4516--4522.
Available via \code{phyloseq::GlobalPatterns}.
}
\usage{
example_microbiome
}
\description{
A subset of the \code{GlobalPatterns} dataset from the \pkg{phyloseq}
package, filtered to five habitat types (Soil, Ocean, Freshwater,
Freshwater creek, Sediment estuary) and reshaped to long format.
Relative abundance (RA) is pre-calculated per ASV per sample.
}
\keyword{datasets}
