% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloPal-package.R
\docType{package}
\name{phyloPal-package}
\alias{phyloPal}
\alias{phyloPal-package}
\title{phyloPal: Taxonomic Color Palettes and Alluvial-Dendrogram Visualization for Microbiome Data}
\description{
Tools for creating publication-ready microbiome visualizations. Provides perceptually uniform HCL color palettes with optional hierarchical grouping by higher taxonomy, taxonomic barplots, alluvial plots for visualizing compositional changes across groups, and combined alluvial-dendrogram layouts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mwslawinska/phyloPal}
  \item Report bugs at \url{https://github.com/mwslawinska/phyloPal/issues}
}

}
\author{
\strong{Maintainer}: Magdalena W. Slawinska \email{mslawinska@mpipz.mpg.de} (\href{https://orcid.org/0000-0002-6938-0270}{ORCID})

}
\keyword{internal}
